[GitHub] Vapor 3 moved models
